Wrap Sidekiq queues in a status object
This commit adds .statuses function, which essentially returns a list of
all Sidekiq queues wrapped in a class. This class adds a bunch of helper
methods, intended to make it easy to query for the health of your queue.

Note that this helper makes a "snapshot" of the time that the status is
constructed. Since the content of a Sidekiq queue changes quite often,
it is necessary to make frequent snapshots for a real-time monitor.